requirements.txt
asgiref==3.7.2
attrs==23.1.0
Django==4.2.7
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.5
inflection==0.5.1
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
psycopg2-binary==2.9.9
PyJWT==2.8.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.0
rpds-py==0.15.2
sqlparse==0.4.4
typing_extensions==4.9.0
uritemplate==4.1.1
